+++ test/api/api_test.cpp
@@ -915,7 +915,7 @@ TEST_F(APITest, globalLimit) {
     std::string value = "var var_";
     value += name;
     value += " = true";
-    this->createTempFile(name.c_str(), value);
+    static_cast<void>(this->createTempFile(name.c_str(), value));
   }
 
   const char *dir = this->getTempDirName();
@@ -941,10 +941,10 @@ source %s/mod_{24572..28666}
   ASSERT_EQ(1, r);
   ASSERT_EQ(AR_ERROR_KIND_TYPE_ERROR, e->kind);
   ASSERT_STREQ("GlobalLimit", e->name);
-  // ASSERT_EQ(1, e->lineNum);
-  // ASSERT_EQ(5, e->chars);
-  ASSERT_EQ(9, e->lineNum);
-  ASSERT_EQ(8, e->chars);
+  ASSERT_EQ(1, e->lineNum);
+  ASSERT_EQ(5, e->chars);
+  // ASSERT_EQ(9, e->lineNum);
+  // ASSERT_EQ(8, e->chars);
   e = newError();
 }